Disable the automatic printing of configuration settings
Due to the large number of configuration options that we utilize, terminals can
be completely filled with just the configuration options when the user executes
ffmpeg or ffprobe. Because of this, and because we know which configuration
options were used in a given copy of Houdini, we can safely disable the line
below. The information can still be found if necessary by passing -buildconf to
the executable.