Reduce the log level of invalid data in MJPEG app fields message
This message is rather common when dealing with Logitech webcams, where
they embed an H.264 stream into the MJPEG stream. In order to avoid
filling the console with warnings, and since FFmpeg doesn't do anything
with this data anyways, we'll reduce the log level such that it's below
the default FFmpeg log level in Houdini.